chore: change data type

+++ migrations/scripts/20241008061631-create-dev-console-user.js
@@ -15,7 +15,7 @@ module.exports = {
                 unique: true,
             },
             balance: {
-                type: Sequelize.FLOAT,
+                type: Sequelize.DECIMAL(20, 5),
                 allowNull: false,
                 defaultValue: 5,
             },
